+++ frontend/src/lib/components/apps/editor/SettingsPanel.svelte
@@ -33,6 +33,17 @@
 							return { item: { data: tableAction, id: tableAction.id }, parent: x.data.id }
 						}
 					}
+				} else if (
+					x?.data?.type === 'aggridcomponent' ||
+					x?.data?.type === 'aggridcomponentee' ||
+					x?.data?.type === 'dbexplorercomponent'
+				) {
+					if (x?.data?.actions) {
+						const tableAction = x.data.actions.find((x) => x.id === id)
+						if (tableAction) {
+							return { item: { data: tableAction, id: tableAction.id }, parent: x.data.id }
+						}
+					}
 				}
 			})
 			.find((x) => x)
@@ -78,6 +89,17 @@
 
 					if (parent.data.type === 'tablecomponent') {
 						parent.data.actionButtons = parent.data.actionButtons.filter(
+							(x) => x.id !== tableActionSettings?.item.id
+						)
+					}
+
+					if (
+						(parent.data.type === 'aggridcomponent' ||
+							parent.data.type === 'aggridcomponentee' ||
+							parent.data.type === 'dbexplorercomponent') &&
+						Array.isArray(parent.data.actions)
+					) {
+						parent.data.actions = parent.data.actions.filter(
 							(x) => x.id !== tableActionSettings?.item.id
 						)
 					}