+++ frontend/src/lib/components/apps/editor/component/Component.svelte
@@ -341,6 +341,7 @@
 				configuration={component.configuration}
 				id={component.id}
 				customCss={component.customCss}
+				actions={component.actions ?? []}
 				bind:initializing
 				{render}
 			/>
@@ -351,6 +352,7 @@
 				bind:initializing
 				componentInput={component.componentInput}
 				customCss={component.customCss}
+				actions={component.actions ?? []}
 				{render}
 			/>
 		{:else if component.type === 'aggridcomponentee'}
@@ -361,6 +363,7 @@
 				bind:initializing
 				componentInput={component.componentInput}
 				customCss={component.customCss}
+				actions={component.actions ?? []}
 				{render}
 			/>
 		{:else if component.type === 'textcomponent'}