+++ frontend/src/routes/(root)/(logged)/flows/add/+page.svelte
@@ -141,7 +141,6 @@
 	on:details={(e) => {
 		goto(`/flows/get/${e.detail}?workspace=${$workspaceStore}`)
 	}}
-	gotoEdit={(path, selectedId) => goto(`/flows/edit/${path}?selected=${selectedId}`)}
 	{initialPath}
 	{pathStoreInit}
 	bind:getSelectedId
@@ -153,8 +152,5 @@
 	{selectedId}
 	{loading}
 >
-	<UnsavedConfirmationModal
-		savedValue={{ ...savedFlow, path: undefined }}
-		modifiedValue={$flowStore}
-	/></FlowBuilder
+	<UnsavedConfirmationModal savedValue={savedFlow} modifiedValue={$flowStore} /></FlowBuilder
 >